Allow inline "delta" flag. Only keep last level for redefined IDL.
To keep the list of specifications readable, the "delta" flag can now be
specified inline next to the URL, separated by a space as in:
 "https://www.w3.org/TR/mydeltaspec/ delta"

This is just syntactic sugar, a real object can still be used:
 {
   "url": "https://www.w3.org/TR/mydeltaspec/",
   "delta": true
 }

The anomalies report did not pay attention to levels when it looked at
redefined IDL terms and as such reported false positives of terms defined
in one level and in the next, which is perfectly normal. We didn't have the
problem before because crawl only contained one level for IDL specs. This case
is now properly handled.