Export latest level of specs that have something to export
CSS/IDL exports are now created for the latest "fullest" level of a spec that
has some something to export (as opposed to just the latest level of a spec).

In other words, the crawl will generate an IDL extract for a spec that is a
delta spec if previous full level does not define any IDL content.

Also dropped a couple of duplicated URLs from the list of specs, and added
unit tests for the isLatestLevelThatPasses function that contains the selection
logic.