Treat all specs the same by not expecting CSS/IDL up front (#231)
Specs provided as input to the crawl used to be separated between specs that
define CSS terms and specs that define IDL terms (with a few specs that appeared
in both lists). This was no longer practical as goal is to switch to a flat list
of specs from https://github.com/w3c/browser-specs/

With this update, the crawler will try to extract CSS and IDL definitions from
all specs. By definition, this update removes the ability to report a warning
when the crawler could not extract CSS/IDL terms from a given spec.

When multiple levels of the same spec are crawled, only the latest level that
is not a delta spec leads to a CSS/IDL extract, unless the delta spec is the
only level that can lead to a CSS/IDL extract for this spec. Whether a spec
is a delta spec needs to be specified in advance with a "delta" flag in the list
of specs.

The "delta" flag can be specified inline next to the URL to keep the list of
URLs readable, as in:
 "https://www.w3.org/TR/mydeltaspec/ delta"

This is just syntactic sugar, a real object can still be used:
 {
   "url": "https://www.w3.org/TR/mydeltaspec/",
   "delta": true
 }

CSS specifications are essentially the only ones that have delta specs for now.

Duplicate URLs have been removed from the list of specs accordingly (spec files
could be merged as well. Not done in this update though as we'll switch to
browser-specs afterwards in any case).

The anomalies report did not pay attention to levels when it looked at redefined
IDL terms and as such reported false positives of terms defined in one level and
in the next, which is perfectly normal. We didn't have the problem before
because crawl only contained one level for IDL specs. This case is now properly
handled.

Additional changes:
* Fix page selector for CSS 2.2 draft, following a recent change in the format
used for the Table of Contents in the spec.
* Fix corner case in references extraction to handle specs that include an empty
"dt" tag (typically the level 4 of the CSS Pages Media Module).
* Unit tests for function that detects whether a spec is the latest level of
a spec that passes some predicate in a list of specs
* Added a few delta CSS specs that were in Shepherd and not in Reffy