Extract dfns that follow Bikeshed's dfn data model
Implement a definition extraction module that follows the rules defined in:
https://tabatkins.github.io/bikeshed/#dfn-contract

With this update, the crawler exports definitions in all specs both in the
`crawl.json` file and in individual JSON files per spec in a `dfns` subfolder.

To correctly compute the URL of the page that contains a particular definition,
the crawler no longer parses single page versions of multi-page specs, but loads
each page one by one (to create a single page version where contents of pages
can be identified).

Here is an example of a definition in the resulting JSON file:

```json
{
  "id": "event-abort",
  "href": "https://html.spec.whatwg.org/multipage/indices.html#event-abort",
  "linkingText": [
    "abort"
  ],
  "localLinkingText": [],
  "type": "event",
  "for": [
    "Window"
  ],
  "exported": true
}
```

The `linkingText` is the text content unless overwritten by a `data-lt`
attribute. It is an array because the attribute can specify multiple texts.

The `for` property lists the namespaces for the definition.

The `exported` property is true when the term can be referenced from external
specifications, false otherwise.