Add parsing of definitions specific to HTML spec (#340)
The HTML spec does not follow the markup conventions to properly type and associate definitions,
so it needs special handling.

A lot is encoded in the ids of the definition elements, but not all the data (e.g. no distinction
between attributes and methods), and there are many exceptions to the general conventions on id
building.

Figuring out the right output requires additional knowledge that is available through the WebIDL
fragments embedded in the spec, so we make use of that.