Generate index file and additional individual extracts
This update partly addresses #385. It updates the code to generate an
`index.json` file as described in the issue, where spec entries only contain
metadata and links to generated dumps.

This update also generates individual extracts that did not exist before, for
the parsed IDL, the links, the headings and the references.