Generate index file and additional individual extracts (#388)
This update partially addresses #385. It updates the code to generate an
`index.json` file as described in the issue, where spec entries only contain
metadata and links to generated dumps.

Additional individual JSON extracts are also generated from the info that was
in `crawl.json` in subfolders that did not exist before:
- `idlparsed` for the parsed IDL
- `links` for the list of links (and fragments) to external documents
- `headings` for the list of headings
-  `refs` for the list of normative and informative references

These JSON files are per spec, named after the spec's shortname, and start with
a short `spec` property that contains the title of the spec and the crawled URL
to make it easier to consume the files without having to go through `index.json`

The study tool no longer uses `crawl.json` but rather uses the `index.json` file
directly. The anomalies report generator tool was restricted to only accept
study files as input, and not also `crawl.json` to generate diffs: the ability
to use `crawl.json` should not really be needed, starting from `study.json`
makes things easier and also avoids having to fetch 27MB of data that would be
only useful to re-create the `study.json` file.

The `crawl.json` file is still generated as before. To be dropped once we're
confident that external projects no longer depend on it.

Note: the `check-specs` CLI is actually broken, see #386. File update in this
commit does not fix that.