Support HTTPS resources in crawl expansion function (#390)
The `expandCrawlResult` utility function was expecting the initial crawl object
to be retreived from the file system. However, it may well have been fetched
over HTTP. The function now allows second parameter to be an HTTPS URL and
fetches resources over HTTP when that is the case.

Note that expanding a typical crawl result is likely going to trigger hundreds
of HTTP requests.