Update tidoust/reffy-reports URLs to target w3c/webref
The tidoust/reffy-reports repository appears in Reffy's code because:
1. reffy.js needs to know where to find the reference study for comparison
2. the HTML templates need to target CSS/JS files that are in the reports
repository.
3. the preamble in IDL extracts links to the repo

This commit updates URLs for 1. (which should fix the ability to crawl) and 2.
(which should fix the generated HTML files).

This commit does not fix 3, because that would update all IDL files, which would
in turn trigger a cascade of dull pull requests onto Web Platform Tests. To be
done later on.