Add a heading property to dfns that target enclosing section
This update adds a `heading` property to dfns extracts. This new property allows
to answer the question: "Where does this definition appear in the spec?".

The `heading` property is an object with 3 keys:
- `id`: either the ID of the section, if there exists an explicit section
element and if that section has an ID, or the ID of the heading element. The ID
may be null is neither the section nor the heading element has an ID, although
that should never happen with "modern" specs since authoring tools automatically
generate these IDs.
- `title`: the title of the closest section's heading element, without the
heading number. E.g. "Interface PresentationReceiver"
- `number`: The heading number, when heading is numbered, e.g. "6.6". Final "."
is ommitted (e.g. "A" and not "A.").

Under the hoods, this update adds the `createOutline` and `mapIdsToHeadings`
functions to Reffy's crawling namespace. The first one computes the proper
outline of the specification, following the "creating an outline" algorithm
defined in HTML. The second one builds on top of the first one to map elements
that have an ID to the logical section that contains them.

This update also modifies the headings extracts to reuse the result of the
`mapIdsToHeadings` function. The net benefit is that the ID of the heading is
now the heading of the section when it's defined, as done in the Table of
Contents. For instance, the ID of the heading "A. IDL Index" in a ReSpec spec
becomes "idl-index" and not "a-idl-index" (which is the ID of the heading
element itself).

(As a side effect, heading's appendix titles are now also correctly parsed).