Add "definedIn" property to dfns extract (#446)
The `definedIn` property captures the context under which the definition appears
in the specification. Value may be one of:
- "dt": appears in a <dt>
- "pre": appears in a <pre>
- "table": appears in a <table>
- "heading": appears in one of <h1>, <h2>, <h3>, <h4>, <h5>, <h6>
- "note": appears in a section flagged with a `note` class
- "example": appears in a section flagged with an `example` class
- "prose": appears elsewhere, typically in a paragraph.

This information is meant to be used for analysis purpose, e.g. to identify
cases where a definition used by a spec is incorrectly defined in a Note, or
where an IDL definition appears in an IDL block instead of in the prose of the
specification.

Other projects may also perhaps use that information, likely in combination with
`type`, to assess whether they should link to the definition itself or to the
section that contains the definition.

The classification may evolve based on feedback.