Add script to generate report on spec links robustness (#455)
Creates a backrefs analyzer command that takes the results of an ED and TR
crawl as inputs and generates a report of problematic references to external
specs. The report contains:

* Links to anchors that do not exist (broken links)
* Links to anchors that exist in the TR version of the targeted spec but no
longer exist in the ED
* Links to anchors that are neither definitions nor headings
* Links to definitions that are not flagged as being exported
* Links to dated TR URLs
* Links to outdated specs (typically those that have now been merged into HTML)
* Links to resources that look like specs but that do not exist in the crawl
and are not well-known informative specs

Note that the tool only analyzes links that seem to target specs. It ignores
links to tests, GitHub issues, Wiki pages, etc.

Returned report uses a Markdown syntax suitable for copy-pasting to a GitHub
issue.