Drop builds folder, generate code dynamically
Starting with v14 (possibly with v12 as well?), Node.js supports loading ES6
modules within CommonJS modules dynamically through the `import()` method. This
makes is possible to get rid of the `canonicalize-url.js` build, provided that
the file gets renamed with a `.mjs` extension. This is what this update does.

This update also renames other ES6 modules with a `.mjs` extension. That is not
really needed, but that seems better for consistency.

With one build that remained, it seemed a good idea to try to get rid of it as
well. This is both harder and easier. Easier because `browser.js` can easily be
generated dynamically through the Rollup API. Harder because that should be done
only once when the program starts, which means that a setup phase needed to be
created. Not fantastic, but hard to avoid.

In turn, that setup phase (along with teardown phase) allows to create the
Puppeteer instance at the beginning and to reuse it throughout the crawl. Which
this update also does.

A new test was added to make sure that building the code is doable. Other tests
were fixed to call setup and teardown functions when needed.

This update means that Node.js >=14 is now needed.

Close #387.
Close #415.