Drop builds folder, generate code dynamically (#458)
Starting with v14 (possibly with v12 as well?), Node.js supports loading ES6
modules within CommonJS modules dynamically through the `import()` method. This
makes is possible to get rid of the `canonicalize-url.js` build, provided that
the file gets renamed with a `.mjs` extension. This is what this update does.

This update also renames other ES6 modules with a `.mjs` extension. That is not
really needed, but that seems better for consistency.

This update also now injects ES6 modules directly in loaded pages, not to have
to bundle browser.js with Rollup first, as discussed in:
https://github.com/w3c/reffy/pull/458#pullrequestreview-550070771

It does so by intercepting network requests to scripts that compose the code
we're trying to inject and responding to them with the contents of local files.

Note Rollup is still used in tests, where the network interception technique
would be a bit heavy.

In turn, the setup phase (along with teardown phase) allows to create the
Puppeteer instance at the beginning and to reuse it throughout the crawl. Which
this update also does. A processSingleSpecification method was added for CLI
commands not to have to worry about the setup/teardown phases, as discussed in:
https://github.com/w3c/reffy/pull/458#issuecomment-743091889

A new test was added to make sure that building the code is doable. Other tests
were fixed to call setup and teardown functions when needed.

This update means that Node.js >=14 is now needed.

Close #387.
Close #415.