Do not consider delta specs for base CSS/IDL extracts
CSS/IDL extracts are created for the latest level in the spec's series that
contains CSS/IDL definitions. On top of that, they are also created for delta
specs.

When a delta spec was the only level in the series definining CSS/IDL, the
crawled created two extracts for it: one named after the series shortname and
one named after the delta shortname, with only the second one being linked from
`index.json`.

This update makes sure that, in such cases, only the extract named after the
delta shortname gets produced.