Integrate dfns and external links analyzes in anomalies reports
This update completes the study tool to integrate recently added checks to the
final JSON report, adding:

- `missingDfns`: an object with `css` and `idl` properties that contain the
results of running the `check-missing-dfns.js` tool.
- `xrefs`: an object with several properties that contain the results of
running the `study-backrefs.js` tool. The final report does not contain the
`unknownSpecs` because they are not anomalies of the spec but rather signals
potential needs to complete browser-specs.

The tools mentioned had to be slightly updated to be integrated in the study
tool.

The `generate-report.js` tool has also been updated to report the new anomalies,
but note that they have not been integrated in the diff reports for now.

Fixes #447.