Do not take for granted that TR report exists (#484)
Following previous update to integrate dfns and backrefs analyses in anomalies
report (see https://github.com/w3c/reffy/pull/482), the reffy.js CLI takes for
granted that the TR report exists when it runs the backrefs analyzer.

Unless one has run the crawler in `tr` mode first, that is not the case. This
update runs the backrefs analyzer without the TR report if it cannot find it.
When that happens, the backrefs analyzer cannot distinguish between
`brokenLinks` and `evolvingLinks`, and reports all such links under
`brokenLinks`.