Fix backrefs analyser crash when spec has no links
If crawling a spec fails for some reason (e.g. a network error), the resulting
crawl report will have no links, and thus no links extract will be created. In
turn, that means that the spec will have no `links` property in `index.json`.

The backrefs analyser expected the `links` property to be set to an array (be it
empty). This update handles the case when the property is not set at all.