Look for node_modules path when tool starts
The location of the `node_modules` folder depends on whether the code is run
directly or as a dependency. When used as a dependency, all dependencies are
flattened to the same parent `node_modules` folder.

The code needs to resolve relative links to `node_modules` because the
`extract-dfns` browser library needs to import the WebIDL parser lib, which is
installed as a dependency.

This update starts by setting a global `node_modules` constant, and uses it
during spec processing to resolve the relative paths appropriately.

Fixes #477.