Add IDL Names generator tool
The IDL names generator takes a crawl report as input and creates a report per
referenceable IDL name, that details the complete parsed IDL structure that
defines the name across all specs.

The parsed IDL structure is a wrapped version of the structure that appears in
`idlparsed` extracts. Here is an example:

```json
{
  "defined": {
    "spec": {
      "title": "Media Capture and Streams",
      "url": "https://www.w3.org/TR/mediacapture-streams/"
    },
    "type": "dictionary",
    "name": "ConstraintSet",
    "inheritance": null,
    "members": [],
    "extAttrs": [],
    "partial": false,
    "href": "https://w3c.github.io/mediacapture-main/#dom-constraintset"
  },
  "extended": [],
  "inheritance": null,
  "includes": []
}
```

The meaning of the properties is:
- `defined` contains the base IDL definition of the name and includes a `spec`
property that describes where the name is defined. Note the URL that appears is
the spec identifier, equivalent to the `url` field in browser-specs, and not
necessarily the crawled URL. The rest of the structure is the `idlparsed` one
(with the exception of the `href` property, see below).
- `extended` contains the list of partial definitions that extend the base
definition, each of them following the same structure as the one presented here.
The order of the list follows the order of appearance in the crawl results,
where specs are sorted by URL.
- `inheritance` contains the inherited interface when there is one, again
following the same structure. The whole inheritance chain appears, meaning that
one can follow `inheritance` properties to get from `HTMLVideoElement` all the
way down to `EventTarget`.
- `includes` contains the list of mixins that the name includes, each of them
following the same structure as the one presented here. The order of the list
follows the order of appearance in the crawl results, where specs are sorted by
URL.

Whenever possible, all IDL terms get linked to their definition in the spec
through an `href` property (which uses the crawled URL). That property is
computed from the dfns extracts. The property appears at the interface level
and also for individual IDL property names, as in:

```json
{
  "defined": {
    "spec": {
      "title": "CSS Spatial Navigation Level 1",
      "url": "https://www.w3.org/TR/css-nav-1/"
    },
    "type": "enum",
    "name": "SpatialNavigationDirection",
    "values": [
      {
        "type": "enum-value",
        "value": "up",
        "href": "https://drafts.csswg.org/css-nav-1/#dom-spatialnavigationdirection-up"
      },
      {
        "type": "enum-value",
        "value": "down",
        "href": "https://drafts.csswg.org/css-nav-1/#dom-spatialnavigationdirection-down"
      },
      {
        "type": "enum-value",
        "value": "left",
        "href": "https://drafts.csswg.org/css-nav-1/#dom-spatialnavigationdirection-left"
      },
      {
        "type": "enum-value",
        "value": "right",
        "href": "https://drafts.csswg.org/css-nav-1/#dom-spatialnavigationdirection-right"
      }
    ],
    "extAttrs": [],
    "href": "https://drafts.csswg.org/css-nav-1/#enumdef-spatialnavigationdirection"
  },
  "extended": [],
  "includes": []
}
```

Related discussion in #472.