Only export raw IDL fragments, not AST
The AST is only half useful because it is not suitable for serialization
purpose. Adding the raw IDL fragment is better from that perspective.

For the export to be readily usable, it needs to include definitions. Without
the AST, these definitions need to be at the root level. In turn, this means
that additional logic is needed when one wants to re-serialize the IDL fragment
to associate the definitions back to the appropriate definitions. That logic is
not fantastically straightforward.

Definitions are not included by default, and not included in the crawl. It would
be good to have clear feedback on whether they are going to be useful.