Add IDL Names generator tool (#489)
The IDL names generator takes a crawl report as input and creates a report per
referenceable IDL name, that details the complete parsed IDL structure that
defines the name across all specs.

The parsed IDL structure is a wrapped version of the structure that appears in
`idlparsed` extracts without the AST structure. Here is an example:

```json
{
  "name": "ConstrainULongRange",
  "defined": {
    "spec": {
      "title": "Media Capture and Streams",
      "url": "https://www.w3.org/TR/mediacapture-streams/"
    },
    "fragment": "dictionary ConstrainULongRange : ULongRange {\n  [Clamp] unsigned long exact;\n  [Clamp] unsigned long ideal;\n};",
    "href": "https://w3c.github.io/mediacapture-main/#dom-constrainulongrange"
  },
  "extended": [],
  "inheritance": {
    "name": "ULongRange",
    "defined": {
      "spec": {
        "title": "Media Capture and Streams",
        "url": "https://www.w3.org/TR/mediacapture-streams/"
      },
      "fragment": "dictionary ULongRange {\n  [Clamp] unsigned long max;\n  [Clamp] unsigned long min;\n};",
      "href": "https://w3c.github.io/mediacapture-main/#dom-ulongrange"
    },
    "extended": [],
    "inheritance": null,
    "includes": []
  },
  "includes": []
}
```

The meaning of the properties is:
- `defined` contains the base IDL definition of the name and includes a `spec`
property that describes where the name is defined. Note the URL that appears is
the spec identifier, equivalent to the `url` field in browser-specs, and not
necessarily the crawled URL. The rest of the structure is the `idlparsed` one
without the AST structure
- `extended` contains the list of partial definitions that extend the base
definition, each of them following the same structure as the one presented here.
The order of the list follows the order of appearance in the crawl results,
where specs are sorted by URL.
- `inheritance` contains the inherited interface when there is one, again
following the same structure. The whole inheritance chain appears, meaning that
one can follow `inheritance` properties to get from `HTMLVideoElement` all the
way down to `EventTarget`.
- `includes` contains the list of mixins that the name includes, each of them
following the same structure as the one presented here. The order of the list
follows the order of appearance in the crawl results, where specs are sorted by
URL.

This update also exports a textual representation of the IDL in a `fragment`
property (added to `idlparsed` exports as well).

Whenever possible, IDL terms get linked to their definition in the spec
through an `href` property (which uses the crawled URL). That property is
computed from the dfns extracts. The property appears at the interface level.

The IDL names generator may also be instructed to add a `dfns` property
that contains all definitions (from dfns extracts) that are relevant to names
that appear in the definition of the IDL. These definitions are not added
by default. They may be useful to serialize IDL fragments with links to
definitions, but this requires custom logic to harness the WebIDL serializer,
see:
https://github.com/w3c/reffy/pull/489#issuecomment-775859515

The crawler also exports one text file per IDL name in an "idlnames" folder.
Each text file contains the full interface (without the fragments that define
the inherited classes).

Related discussion in #472.