Don't clobber previously identified definitions (#500)
The CSS definitions extractor may find and parse multiple definitions for the
same term (e.g. one definition in a `<pre>`, another in a `<dl>`, etc.),
extracting part of the definition each time.

This update makes sure that the extractor preserves partial info it already has
on a term when it finds additional information.