Add crawling tests using Reffy as a module
This update adds a test similar to the tests in `crawl.js` that makes sure
that the crawler runs successfully when Reffy is installed as a module.

The test packs Reffy using `npm pack`, installs it to a tmp folder, and then
runs the crawler in that folder.

Common logic between `crawl.js` and the new test is now defined in a
`nock-server.js` lib file.