Look for node_modules path when tool starts (#487)
The location of the `node_modules` folder depends on whether the code is run
directly or as a dependency. When used as a dependency, all dependencies are
flattened to the same parent `node_modules` folder.

The code needs to resolve relative links to `node_modules` because the
`extract-dfns` browser library needs to import the WebIDL parser lib, which is
installed as a dependency.

This update starts by setting a global `node_modules` constant, and uses it
during spec processing to resolve the relative paths appropriately.

This update also adds a test similar to the tests in `crawl.js` that makes sure
that the crawler runs successfully when Reffy is installed as a module.

The test packs Reffy using `npm pack`, installs it to a tmp folder, and then
runs the crawler in that folder.

Common logic between `crawl.js` and the new test is now defined in a
`nock-server.js` lib file.

Fixes #477.