Fix name for undefined.idl files
The code could produce files named `undefined.idl` in the `idlnames` folder when
it bumped on a partial definition for which the base definition does not exist
in the crawl.

Ideally, these situations should never happen. They do in practice when a spec
cannot be crawled e.g. for transient network issues or because it contains
invalid IDL.

Note that the code was already correct for the `idlnamesparsed` folder.