CSS valuespace prose dfn: stick to the core definition
CSS valuespaces defined in `<dd>` are typically defined using prose. The crawler
was taking the whole block as definition, converting it to raw text. This worked
well in most cases, but this block sometimes contain additional information that
cannot easily be converted to raw text and that goes too much into details for
the purpose at hand, namely:
- notes
- links to tests
- child `<dl>` lists that details possible values

This update excludes these bits from the extracted prose.

Fixes #598