Extract CSS dfns only in normative sections
css-overflow-4 currently defines a possible extension to scrollbars-gutter in an
informative section. That definition was extracted and should not be. This
update restricts CSS dfns extraction to normative sections and fixes #641.

This update also detects situations where a CSS property is defined more than
once in the same spec. The code now throws when that happens instead of
overriding the first definition silently. We may want to improve the handling
of such situations in the future and e.g. report them as an array and then as
an anomaly to fix.

This update also fixes #642, returning descriptors in an array indexed by
descriptor name, instead of expecting descriptor names to always be unique
within a single spec. This change is breaking but cannot be avoided as there
are and will be two `font-display` descriptors in the CSS Fonts spec: one
for the `@font-face` rule and one for the `@font-feature-values` rule.