Filter out previous spec versions when looking for IDL names (#649)
The code that generated the list of IDL names extracted names from all specs.
IDL names may be defined multiple times in a given series though (e.g. in
level 1, 2 and 3). The info appeared duplicated as a result.

This update applies the same rules as the ones used to produce the IDL extracts:
only look at the latest "fullest" version of a spec that defines some IDL,
completed with delta specs.

This fixes #648 and also fixes duplication in `Performance.idl` that was due to
`user-timing-2` and `user-timing-3`.