Use absolute URLs for dfn headings, headings, and ids
Multipage specs were badly supported because the code assumed that local IDs
would be unique for the entire document, but that is generally speaking not
the case.

This update adds an `href` property with an absolute URL throughout extracts.
That was already the case for dfns. This is now also done for headings in dfn
extracts and for the headings extracts themselves.

The ids extracts now also use absolute URLs everywhere.

Local IDs still appear in the `id` property.

Fixes #647.