Ignore dfns that have an invalid type
This update restricts dfns extraction to definitions that have a valid type,
as defined in:
https://tabatkins.github.io/bikeshed/#dfn-types
(+ `namespace` and `event` that are valid but do not yet appear in the doc)

In practice, the only invalid type that appears in Editor's Drafts is `idl`,
which ReSpec uses to flag internal slots. To avoid losing definitions while
ReSpec gets fixed and while the changes propagate to all specs and /TR versions
of the spec, the code automatically converts `idl` definition types to
`attribute`.

Data validation is something that we would typically do at a later stage,
through anomaly reports, or through patches as we produce packages. That said,
dfns are more time sensitive than other types of data that Reffy extracts from
the specs and the validation process would need to be semi-manual.

Dfns with an invalid type get reported as warnings to the console.

Fixes #658.