Generalize crawl-specs to accept mjs modules as input (#694)
This update changes the processing logic to take a list of ES6 processing
modules as input, allowing to turn the specs crawler into a generic tool that
can be used to extract custom information from Web specs.

This update completely re-writes the crawl-specs CLI tool in particular, turning
the CLI into a "proper" command with options, and extending options to add a
"--module" option that lets user target custom processing modules, as needed.

In the absence of CLI options, the crawler behaves exactly as before, crawling
all browser-specs and running all core processing modules to create the usual
crawl report. The new "--module" CLI option allows users to take control of the
processing bits.

Run "node src/cli/crawl-specs.js -h" for detailed usage information.

Under the hoods, the crawler creates what used to be `reffy.mjs` on-the-fly from
the list of modules given as input. That file is thus no longer needed.

The command has provisions to reference core processing modules if needed.

The logic is still slightly clunky here and there. Handling of relative paths
to modules remains hacky for instance. Some further updates that could be
envisioned:
- Re-write other tools that use `processSingleSpecification` to go through the
crawler instead with the right module.
- Make it possible to run custom processing modules that create one extract per
spec series (using `isLatestLevelThatPasses`) instead of one extract per spec.
- Reduce the need to maintain dedicated logic for the CSS and IDL processing
modules.
- Validate CLI arguments more thoroughly (e.g. check that processing modules
actually exist, that a given processing module appears only once, etc.)
- Write tests against the CLI tools.

Some side updates:
- Processing modules now always receive the spec object as first parameter.
- Specs crawler was slightly improved to allow users to use an arbitrary spec
URL in the list of specs to crawl. This is one of the things that remain clunky
for now, as the crawler would happily make use of a spec shortname if it had
one, and currently merely compacts the URL to produce a shortname.