Fix isEmpty logic, replace newlines in titles
The update makes sure that browserlib modules no longer produce weird empty
constructs such as objects with keys at the first level but without real values.

The isEmpty logic was updated, first to fix the array case (`typeof` never
returns `array` because because), and second to remain as dumb as possible.

This gave a good opportunity to get rid of module defaults which are indeed
completely useless as envisioned by the TODO comment.

Last update drops extraneous spaces in extracted titles to avoid them generating
separate paragraphs in the final report.

Fixes #716 and #717.