Improve crawler CLI and get rid of individual processing CLI tools
Continued to rewamp the crawler CLI:
- Turned the report folder CLI parameter into an `--output` option. If not
provided, the crawler reports results to the console.
- Added a `--quiet` CLI option to the crawler CLI not to report warnings and
progress messages. Also made crawler output internal messages as warnings so
that stdout remains clean for the final result.
- Added a `--spec` CLI option to pass spec URLs or shortnames directly on the
command-line instead of having to put them in a JSON file.
- Added a `--terse` CLI option to return the results of the processing directly
instead of wrapping it into an array structure. The option cannot be combined
with `--output` and can only be set if only one processing module runs

The above changes made it possible to drop the `processSingleSpecification`
logic and individual processing CLIs. This can now be achieved with the right
CLI options. For instance `crawl-specs -s [shortname] -m css -t` yields the same
results as the previous `extract-cssdfns` CLI.

The crawling logic was also improved to save extracts as soon as processing on
the spec is done (as opposed to waiting until the end of the crawl). This makes
it possible to run processing modules that create large extracts per spec
without having to worry too much about memory. This fixes #720.

Custom logic for CSS/IDL that generates extracts per series instead of extracts
per spec is now done at the end of the crawl. It renames file extracts
accordingly.

Similarly, the generation of the IDL names extracts and index gets done at the
end of the crawl when needed.

Other changes:
- The `parse-webidl` CLI now takes an IDL file instead of a URL. IDL extraction
can be achieved with `crawl-specs -s shortname -m idl > res.idl` and `res.idl`
can then be passed to `parse-webidl`.
- README file adjusted (and `crawl-specs -h` completed)
- Bumped version to 3.0.0