Add possibility to crawl "file" URLs
The crawl only supported http and https URLs out of the box, but that constraint
was more artificial than anything else. Moving the check on the scheme to the
appropriate place is enough to enable support of "file://" URLs in practice.

This is useful to run the crawler on a local spec and typically needed to
integrate IDL validation in spec-prod.

The crawler CLI now also interprets relative filenames as "file://" URLs,
making it possible to run the crawler on a local file with commands such as:

node src/cli/crawl-specs.js -s spec.html