Fix minimal usage example, prevent CLI from running without option
The minimal usage example was outdated and now incorrect. This fixes it to
`reffy -o .` with a short description of what that means.

By default, the CLI crawls all specs (that's fine), runs all processing modules
(that's fine too) but dumps the results to the console, which is a good default
in theory be it only for the fact that the crawler is going to produce ~100MB of
data if it runs all processing modules on all specs...

This update makes it impossible to run reffy without providing at least one of
the `--output`, `--module` or `--spec` option. If one really wants to dump
things to the console, well, that remains doable by running `reffy -s all`.