Increase max relative depth supported for custom modules (#752)
The code needs to create some deep-enough path to resolve paths to custom
processing modules. The code assumed that 8 would be a good depth, but it
turns out it's easy to need something deeper. This update goes to 11^W 20.

There may be a better way to achieve the same thing and get rid of this
artificial maximum somehow. To be continued...