Create cache folder in install folder, not run folder
The local cache folder (`.cache`) was created where the command was run. If you
install Reffy globally, you won't expect it to create files in the folder you
run it into by default.

This update creates the `.cache` folder in the install folder by default. It may
not be good practice to put a cache that will typically contain dozens of MB of
data in there, but that seems better than putting a cache folder where ever the
command is run in any case.

Note one can override the default through the `cacheFolder` setting in a
`config.json` file.