Create cache folder in install folder, not run folder (#756)
The local cache folder (`.cache`) was created where the command was run. If you
install Reffy globally, you won't expect it to create files in the folder you
run it into by default.

This update uses a cache folder in the OS temp folder.

Note one can override the default through the `cacheFolder` setting in a
`config.json` file.